"Missing privilege separation directory: /run/sshd" on creating user
Unlike for Debian 9, the privilege separation directory /run/sshd is
not automatically created for Debian 10. The configd_update script for
sshd does create this, so if sshd is enabled. But vyatta-update-ssh.pl
may also be called if sshd is not enabled and due to changes in system
login, if only to update the ClientAliveInterval parameter based upon
session-timeout. The more recently added functions for setting up
ciphers and kexalgorithms call sshd, which complains if /run/sshd does
not exist.

One option is not to generate ciphers and kexalgorithms if sshd is not
enabled. The solution chosen though is for the script simply to create
/run/sshd if it does not exist.

Also remove the deprecated UsePrivilegeSeparation parameter from the
generated sshd_config so as to avoid errors in the journal for this.