Commit failure with ssh cipher options cast128-cbc and blowfish-cbc
The CBC ciphers are considered insecure and are not used by default in
openssh. Due to operational requirements, these legacy ciphers had to
be allowed via a 'set service ssh permit' CLI introduced for this. Now
openssh no longer even accepts cast128-cbc and blowfish-cbc, so render
these obsolete. It has been confirmed that the initial need for CBC
ciphers was too generously interpreted, as the operational requirement
is that only the aes128-cbc, aes192-cbc and aes256-cbc legacy ciphers
are still needed.

VRVDR-50501