Remove "legacy" cipher 3des-cbc as option for SSH
The CBC ciphers are considered insecure and are not used by default in
openssh. Due to operational requirements, these legacy ciphers had to
be allowed via a 'set service ssh permit' CLI introduced for this.
NIST has now declared the usage of 3DES as deprecated, so render use
of 3des-cbc obsolete. It has been confirmed that the initial need for
the reinstatement of CBC ciphers was too generously interpreted, as
the operational requirement is that only the aes128-cbc, aes192-cbc
and aes256-cbc legacy ciphers are still needed.

VRVDR-50535