Address NCS-specific YANG compile error
Part of VRVDR-45130.

Original YANG change for this is fine, except that the ancient version of the Cisco NCS used by Flexware struggles with must statement syntax of the format 'function1() * function2()' as it doesn't deal with a function immediately following the '*' operator.  The trivially simple fix is to enclose 'function2()' (count in this case) in brackets, as while equivalent in terms of result, it causes the evaluation order to change, and this is enough to stop the NCS from complaining.

Tested on the NCS.