ssh for ipv6 fails upon reboot
If ssh is configured to listen on an IPv6 address that is not assigned,
then bind to this address fails. This occurs if the address is not
configured or is in tentative state. The latter is the case if the
interface it is on is still down or if DAD has not yet completed after
the interface comes up. As sshd does not support non-local bind and
setting this globally is not an option, a solution is to store listen
addresses so that these can be checked when they do become assigned.

VRVDR-54478