Fix creation of sshd_config for Debian11
Addresses VRVDR-56290.

vyatta-update-ssh.pl adds a 'kexalgorithms' line to sshd_config. It does this by running '/usr/bin/sshd -T -f /dev/null | grep kexalgorithm' to get the supported algorithms and then outputs the list to sshd_config. This no longer works in Debian 11 as there is a new 'gssapikexalgorithms' entry in the output from /usr/bin/sshd as well as the existing 'kexalgorithms' one. This means that the text added to sshd_config is malformed and sshd fails to start.

The proposed fix just filters out the 'gssapikexalgorithms' line by anchoring the grep search to the start of the line.

With this change, the Debian 11 ISO for Luton (only available privately for now) starts sshd successfully.